// src/types/categories/ICategory.ts
import { type IImage } from '@/types/utils/IImage'
import { type IItemBorderInfo } from '@/types/utils/IItemBorderInfo'

export interface ICategory {
  alt: { main: string }
  autogeneratedSlug: string
  borderInfo: IItemBorderInfo
  customSlug: string
  description: string
  descriptionTranslated: {
    [key: string]: string
  }
  enabled: boolean
  hdThumbnailUrl: string
  id: number
  imageUrl: string
  isSampleCategory: boolean
  name: string
  nameTranslated: {
    [key: string]: string
  }
  orderBy: number
  originalImage: IImage
  originalImageUrl: string
  productCount: number
  seoDescription: string
  seoTitle: string
  thumbnail: IImage
  thumbnailUrl: string
  url: string
}